back/pyproject.toml
[project]
name = "python-uv-ruff-template"
version = "0.1.0"
description = "FastAPI application"
readme = "README.md"
requires-python = "==3.12.6"
dependencies = [
    "fastapi>=0.110.0",
    "google-cloud-aiplatform>=1.81.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.29.0",
    "vertexai>=1.43.0",
]

[project.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
lint-imports = "import-linter lint"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "types-requests>=2.32.0.20241016",
    "import-linter>=2.2",
]

[tool.ruff]

# 1行の最大文字数を120に設定
line-length = 120

# インデント
indent-width = 4

# Lint対象外ディレクトリ
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
]

[tool.ruff.format]
# docstring内のコードブロックをフォーマットする
docstring-code-format = true

[tool.ruff.lint]
# すべてのルールを有効化
select = ["ALL"]

# 以下のルールは無視する
# T201: print文の使用を禁止するルール
# COM812: トレーリングカンマに関するルール
# ISC001: 暗黙的文字列連結に関するルール
# D104: パブリックパッケージにdocstring必須に関するエラー
ignore = [
    "T201",
    "COM812",
    "ISC001",
    "D104",
    "D100",
    "INP001",
    "B904", 
    "D415",
]

# 自動修正を適用しないルール
# F401: 未使用のインポート
# F841: 未使用の変数
unfixable = [
    "F401",
    "F841",
]

# docstringのスタイルをGoogle形式に設定
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
# テストファイルに対する特別なルール除外設定
# D: ドキュメント関連のすべてのルール
# S101: assertの使用を許可
# N802: 関数名の命名規則
# ARG: 関数の引数に関するルール
"tests/**/*.py" = [
    "D",
    "S101",
    "N802",
    "ARG",
]

[tool.ruff.lint.pylint]
# 関数の最大引数数を6個に制限
max-args = 6

[tool.pytest.ini_options]
pythonpath = "app"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
mypy_path = "app"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true
